+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - refactor(core): migrate `ubo_app/services.py` to `typings/ubo_handle.pyi` as it was only providing types
 - fix(core): make sure app exits gracefully before shutdown/reboot and atexit callbacks run - closes #150
 - refactor(core): kill the app (so that it restarts) when the reset button is pressed for 0.5 seconds or more and reboot the device when it is pressed for 3 seconds or more - closes #116
+- fix(vscode): after login with vscode, the gui now goes back to login code page - closes #143
 
 ## Version 0.15.5